rutas mejoradas

+++ app/Http/Controllers/RegisterController.php
@@ -58,6 +58,11 @@ class RegisterController extends Controller
     public function delete(Request $request)
     {
         try {
+            //la API no puede quedar sin administradores
+            if(User::count() === 1){
+                return jsonResponse(message: "La API solo tiene un administrador", status: 403, errors: ["error" => "No se puede borrar el administrador"]);
+            }
+
             $userData = User::find($request->id);
             $userData->delete();
             return jsonResponse(data: ["user_id" => $request->id], message: "User deleted", status: 201);
